added smoke test - added BeanConfigSmokeTest - added AbstractTest and AbstractSpringBootTest annotations - added common package in test module - moved AbstractMockMvcTest and utils package to common package
