+++ src/test/java/com/example/universe/simulator/entityservice/common/AbstractMockMvcTest.java
@@ -1,4 +1,4 @@
-package com.example.universe.simulator.entityservice;
+package com.example.universe.simulator.entityservice.common;
 
 import com.example.universe.simulator.entityservice.exception.ErrorCodeType;
 import com.example.universe.simulator.entityservice.exception.RestErrorResponse;
@@ -7,7 +7,6 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.mock.web.MockHttpServletResponse;
-import org.springframework.test.context.ActiveProfiles;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.RequestBuilder;
@@ -17,7 +16,6 @@ import java.io.UnsupportedEncodingException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.asyncDispatch;
 
-@ActiveProfiles("test")
 public abstract class AbstractMockMvcTest {
 
     @Autowired
