+++ src/test/java/com/example/universe/simulator/entityservice/common/AbstractSpringBootTest.java
@@ -1,0 +1,16 @@
+package com.example.universe.simulator.entityservice.common;
+
+import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import org.springframework.boot.test.context.SpringBootTest;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@AbstractTest
+@SpringBootTest
+@AutoConfigureTestDatabase
+public @interface AbstractSpringBootTest {}
