+++ src/test/java/com/example/universe/simulator/entityservice/common/AbstractTest.java
@@ -1,0 +1,13 @@
+package com.example.universe.simulator.entityservice.common;
+
+import org.springframework.test.context.ActiveProfiles;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.TYPE)
+@Retention(RetentionPolicy.RUNTIME)
+@ActiveProfiles("test")
+public @interface AbstractTest {}
