+++ src/test/java/com/example/universe/simulator/entityservice/integration/AbstractIntegrationTest.java
@@ -1,11 +1,9 @@
 package com.example.universe.simulator.entityservice.integration;
 
-import com.example.universe.simulator.entityservice.AbstractMockMvcTest;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
+import com.example.universe.simulator.entityservice.common.AbstractMockMvcTest;
+import com.example.universe.simulator.entityservice.common.AbstractSpringBootTest;
 import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
-import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+@AbstractSpringBootTest
 @AutoConfigureMockMvc
-@AutoConfigureTestDatabase
 abstract class AbstractIntegrationTest extends AbstractMockMvcTest {}
