+++ src/test/java/com/example/universe/simulator/entityservice/integration/GalaxyIntegrationTest.java
@@ -3,8 +3,8 @@ package com.example.universe.simulator.entityservice.integration;
 import com.example.universe.simulator.entityservice.dtos.GalaxyDto;
 import com.example.universe.simulator.entityservice.exception.ErrorCodeType;
 import com.example.universe.simulator.entityservice.filters.GalaxyFilter;
-import com.example.universe.simulator.entityservice.utils.JsonPage;
-import com.example.universe.simulator.entityservice.utils.TestUtils;
+import com.example.universe.simulator.entityservice.common.utils.JsonPage;
+import com.example.universe.simulator.entityservice.common.utils.TestUtils;
 import com.fasterxml.jackson.core.type.TypeReference;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
