+++ src/test/java/com/example/universe/simulator/entityservice/integration/MoonIntegrationTest.java
@@ -6,8 +6,8 @@ import com.example.universe.simulator.entityservice.dtos.PlanetDto;
 import com.example.universe.simulator.entityservice.dtos.StarDto;
 import com.example.universe.simulator.entityservice.exception.ErrorCodeType;
 import com.example.universe.simulator.entityservice.filters.MoonFilter;
-import com.example.universe.simulator.entityservice.utils.JsonPage;
-import com.example.universe.simulator.entityservice.utils.TestUtils;
+import com.example.universe.simulator.entityservice.common.utils.JsonPage;
+import com.example.universe.simulator.entityservice.common.utils.TestUtils;
 import com.fasterxml.jackson.core.type.TypeReference;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
