+++ src/test/java/com/example/universe/simulator/entityservice/integration/StarIntegrationTest.java
@@ -4,8 +4,8 @@ import com.example.universe.simulator.entityservice.dtos.GalaxyDto;
 import com.example.universe.simulator.entityservice.dtos.StarDto;
 import com.example.universe.simulator.entityservice.exception.ErrorCodeType;
 import com.example.universe.simulator.entityservice.filters.StarFilter;
-import com.example.universe.simulator.entityservice.utils.JsonPage;
-import com.example.universe.simulator.entityservice.utils.TestUtils;
+import com.example.universe.simulator.entityservice.common.utils.JsonPage;
+import com.example.universe.simulator.entityservice.common.utils.TestUtils;
 import com.fasterxml.jackson.core.type.TypeReference;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.MediaType;
